+++ src/Core/DecimalComparison.h
@@ -1,14 +1,13 @@
 #pragma once
 
-#include <base/arithmeticOverflow.h>
-#include <Core/DecimalFunctions.h>
-#include <DataTypes/DataTypesNumber.h>
-#include <DataTypes/DataTypesDecimal.h>
+#include <Columns/ColumnConst.h>
 #include <Columns/ColumnVector.h>
 #include <Columns/ColumnsNumber.h>
-#include <Columns/ColumnConst.h>
-#include <Functions/FunctionHelpers.h>  /// TODO Core should not depend on Functions
-
+#include <Core/ColumnsWithTypeAndName.h>
+#include <Core/DecimalFunctions.h>
+#include <DataTypes/DataTypesDecimal.h>
+#include <DataTypes/DataTypesNumber.h>
+#include <base/arithmeticOverflow.h>
 
 namespace DB
 {
@@ -171,7 +170,7 @@ private:
             A a = c0_const.template getValue<A>();
             B b = c1_const.template getValue<B>();
             UInt8 res = apply<check_overflow, scale_left, scale_right>(a, b, scale);
-            return DataTypeUInt8().createColumnConst(c0->size(), toField(res));
+            return DataTypeUInt8().createColumnConst(c0->size(), Field(res));
         }
 
         ColumnUInt8::Container & vec_res = c_res->getData();
