+++ src/Dictionaries/RangeHashedDictionary.h
@@ -29,7 +29,6 @@
 
 #include <Columns/ColumnNullable.h>
 
-#include <Functions/FunctionHelpers.h>
 #include <Interpreters/castColumn.h>
 
 namespace DB
